fix: before_app_first_request deprecation
