+++ invenio_oauthclient/views/client.py
@@ -39,28 +39,6 @@ rest_blueprint = Blueprint(
 )
 
 
-@blueprint.record_once
-def post_ext_init(state):
-    """Setup blueprint."""
-    app = state.app
-
-    app.config.setdefault(
-        "OAUTHCLIENT_SITENAME", app.config.get("THEME_SITENAME", "Invenio")
-    )
-    app.config.setdefault(
-        "OAUTHCLIENT_BASE_TEMPLATE",
-        app.config.get("BASE_TEMPLATE", "invenio_oauthclient/base.html"),
-    )
-    app.config.setdefault(
-        "OAUTHCLIENT_COVER_TEMPLATE",
-        app.config.get("COVER_TEMPLATE", "invenio_oauthclient/base_cover.html"),
-    )
-    app.config.setdefault(
-        "OAUTHCLIENT_SETTINGS_TEMPLATE",
-        app.config.get("SETTINGS_TEMPLATE", "invenio_oauthclient/settings/base.html"),
-    )
-
-
 @blueprint.route("/login")
 def auto_redirect_login(*args, **kwargs):
     """Handles automatic redirect to external auth service.
