+++ setup.cfg
@@ -86,7 +86,7 @@ invenio_base.secret_key =
 invenio_i18n.translations =
     invenio_oauthclient = invenio_oauthclient
 invenio_base.finalize_app =
-    invenio_oauthclient = invenio_oauthclient.finalize_app:finalize_app
+    invenio_oauthclient = invenio_oauthclient.ext:finalize_app
 
 [build_sphinx]
 source-dir = docs/
